+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 * I bumped the major version for the new year.
 
+* I added GetPath and GetValue extension methods for the IConfiguration type.
+
+* I added the ConfigurationException type back again.
+
 ## 2020.6
 
 * I moved to an azure CI/CD pipeline.